+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 ## 1.14.1 (Unreleased)
+
+BUG FIXES:
+
+* provider: Fix the provider about `insecure` didn't take effect[GH-53]
+
 ## 1.14.0 (October 13, 2019)
 
 FEATURES: